+++ test/mocking_runner.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/git-town/git-town/v8/src/subshell"
+	"github.com/git-town/git-town/v8/test/envvars"
 	"github.com/kballard/go-shellquote"
 )
 
@@ -179,18 +180,18 @@ func (r *MockingRunner) RunWith(opts *Options, cmd string, args ...string) (stri
 		opts.Env = os.Environ()
 	}
 	// set HOME to the given global directory so that Git puts the global configuration there.
-	opts.Env = ReplaceEnvVar(opts.Env, "HOME", r.homeDir)
+	opts.Env = envvars.Replace(opts.Env, "HOME", r.homeDir)
 	// add the custom origin
 	if r.testOrigin != "" {
-		opts.Env = ReplaceEnvVar(opts.Env, "GIT_TOWN_REMOTE", r.testOrigin)
+		opts.Env = envvars.Replace(opts.Env, "GIT_TOWN_REMOTE", r.testOrigin)
 	}
 	// add the custom bin dir to the PATH
 	if r.usesBinDir {
-		opts.Env = PrependEnvPath(opts.Env, r.binDir)
+		opts.Env = envvars.PrependPath(opts.Env, r.binDir)
 	}
 	// add the custom GIT_EDITOR
 	if r.gitEditor != "" {
-		opts.Env = ReplaceEnvVar(opts.Env, "GIT_EDITOR", filepath.Join(r.binDir, "git_editor"))
+		opts.Env = envvars.Replace(opts.Env, "GIT_EDITOR", filepath.Join(r.binDir, "git_editor"))
 	}
 	// set the working dir
 	opts.Dir = filepath.Join(r.workingDir, opts.Dir)
